+++ src/main/java/icu/samnyan/aqua/sega/diva/model/request/ingame/StageResultRequest.java
@@ -14,7 +14,7 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class StageResultRequest extends BaseRequest {
-    public int pd_id;
+    public long pd_id;
     public int accept_idx;
     public int start_idx;
     public int hp_vol;
