+++ src/main/java/icu/samnyan/aqua/sega/diva/model/response/card/CardProcedureResponse.java
@@ -15,7 +15,7 @@ public class CardProcedureResponse extends BaseResponse {
     private Result cd_adm_result;
     private int chg_name_price = 100;
     private int accept_idx;
-    private int pd_id;
+    private long pd_id;
     private String player_name;
     private Integer lv_num;
     private Integer lv_pnt;
@@ -25,7 +25,7 @@ public class CardProcedureResponse extends BaseResponse {
     private Integer vcld_pts;
     private PassStat passwd_stat;
 
-    public CardProcedureResponse(String cmd, String req_id, String stat, Result cd_adm_result, int chg_name_price, int accept_idx, int pd_id, String player_name, Integer lv_num, Integer lv_pnt, String lv_str, Integer lv_efct_id, Integer lv_plt_id, Integer vcld_pts, PassStat passwd_stat) {
+    public CardProcedureResponse(String cmd, String req_id, String stat, Result cd_adm_result, int chg_name_price, int accept_idx, long pd_id, String player_name, Integer lv_num, Integer lv_pnt, String lv_str, Integer lv_efct_id, Integer lv_plt_id, Integer vcld_pts, PassStat passwd_stat) {
         super(cmd, req_id, stat);
         this.cd_adm_result = cd_adm_result;
         this.chg_name_price = chg_name_price;
