+++ src/main/java/icu/samnyan/aqua/sega/diva/model/response/card/ChangePasswdResponse.java
@@ -13,9 +13,9 @@ import lombok.Setter;
 public class ChangePasswdResponse extends BaseResponse {
     private Result cd_adm_result;
     private int accept_idx;
-    private int pd_id;
+    private long pd_id;
 
-    public ChangePasswdResponse(String cmd, String req_id, String stat, Result cd_adm_result, int accept_idx, int pd_id) {
+    public ChangePasswdResponse(String cmd, String req_id, String stat, Result cd_adm_result, int accept_idx, long pd_id) {
         super(cmd, req_id, stat);
         this.cd_adm_result = cd_adm_result;
         this.accept_idx = accept_idx;
